Added menu asset loading mechanism to game state
